Task environment changes, use resolved store path
Previously when evaluating tasks and passing flake URLs to tasks
we would just pass the repo path on disk, but this meant that if
the repo changes on disk while tasks are being executed, there's
a risk of things breaking.

Now we resolve the repo flake to a /nix/store/*-source path and then
use that as the flake path for further invocations and when passing
the path of the flake to tasks themselves (e.g for building
NixOS systems).

We still pass the original repo path when using nix-task shell as
it's actually desired behaviour for updates to be taken into account
while nix-task shell is running.